+++ App/sample/commands/context-menu/user.js
@@ -4,12 +4,12 @@ module.exports = {
     category: 'context-menu',
     data: new ContextMenuCommandBuilder()
         .setName('User')
-        .setType(ApplicationCommandType.User),
+        .setType(ApplicationCommandType.User)
+        .setDMPermission(false),
     async execute(interaction) {
         const user = interaction.targetUser;
         const member = interaction.targetMember;
         const embed = new EmbedBuilder()
-            .setColor(0xFFFFFF)
             .setThumbnail(`${user.displayAvatarURL()}`)
             .addFields(
                 { name: 'User:', value: `${user.username}`, inline: true },