+++ App/src/commands/utility/bot-check.js
@@ -14,16 +14,12 @@ module.exports = {
         const user = interaction.options.getUser('user') || interaction.user;
         let embed, message;
         if (user.bot) {
-            embed = new EmbedBuilder()
-                .setAuthor({ name: `${user.tag} is a BOT!`, iconURL: `${user.displayAvatarURL({ dynamic: true, size: 512 })}` })
-                .setColor(0xFFFFFF);
+            embed = new EmbedBuilder().setAuthor({ name: `${user.tag} is a BOT!`, iconURL: `${user.displayAvatarURL({ dynamic: true, size: 512 })}` });
             message = await interaction.reply({ embeds: [embed], fetchReply: true });
             if (message.channel) message.react('🤖');
         }
         else {
-            embed = new EmbedBuilder()
-                .setAuthor({ name: `${user.tag} is not a BOT.`, iconURL: `${user.displayAvatarURL({ dynamic: true, size: 512 })}` })
-                .setColor(0xFFFFFF);
+            embed = new EmbedBuilder().setAuthor({ name: `${user.tag} is not a BOT.`, iconURL: `${user.displayAvatarURL({ dynamic: true, size: 512 })}` });
             message = await interaction.reply({ embeds: [embed], fetchReply: true });
             if (message.channel) message.react('👤');
         }